refactor: extract quality event stats accumulators
