+++ src/stats/calculators/mod.rs
@@ -69,6 +69,8 @@ pub mod bump_stats;
 pub use bump_stats::*;
 pub mod ceiling_shot;
 pub use ceiling_shot::*;
+pub mod ceiling_shot_stats;
+pub use ceiling_shot_stats::*;
 pub mod center;
 pub use center::*;
 pub mod center_stats;
@@ -97,6 +99,8 @@ pub mod goal_tags;
 pub use goal_tags::*;
 pub mod half_flip;
 pub use half_flip::*;
+pub mod half_flip_stats;
+pub use half_flip_stats::*;
 pub mod half_volley;
 pub use half_volley::*;
 pub mod match_stats;
@@ -133,6 +137,8 @@ pub mod settings;
 pub use settings::*;
 pub mod speed_flip;
 pub use speed_flip::*;
+pub mod speed_flip_stats;
+pub use speed_flip_stats::*;
 pub mod territorial_pressure;
 pub use territorial_pressure::*;
 pub mod touch;
@@ -145,6 +151,8 @@ pub mod wall_aerial_shot;
 pub use wall_aerial_shot::*;
 pub mod wavedash;
 pub use wavedash::*;
+pub mod wavedash_stats;
+pub use wavedash_stats::*;
 pub mod whiff;
 pub use whiff::*;
